+++ src/funding/mybank/config.jsx
@@ -11,6 +11,8 @@ export function getMybankConfig() : FundingSourceConfig {
         ...DEFAULT_FUNDING_CONFIG,
 
         defaultLabel: BUTTON_LABEL.MYBANK,
+
+        shippingChange: false,
     
         layouts: [
             BUTTON_LAYOUT.VERTICAL
@@ -20,7 +22,7 @@ export function getMybankConfig() : FundingSourceConfig {
             [ BUTTON_LABEL.MYBANK ]: {
                 ...DEFAULT_LABEL_CONFIG,
     
-                Label: MybankLogo
+                Label: ({ logoColor, optional }) => MybankLogo({ logoColor, optional })
             }
         }
     };