+++ src/interface/marks.js
@@ -1,0 +1,3 @@
+/* @flow */
+
+export { Marks } from '../marks/component';